adding api call to workdistribution
