+++ src/coreModule/teamAdminPanel/WorkDistribution/WorkGroup.jsx
@@ -1,27 +1,54 @@
 import {
-  BsBook,
-  BsBriefcase,
-  BsBuilding,
-  BsCalendar,
   BsCalendar2Event,
   BsCardText,
   BsClipboardCheck,
-  BsCreditCard2Front,
-  BsCurrencyDollar,
-  BsDroplet,
-  BsEnvelope,
-  BsGenderAmbiguous,
-  BsHouse,
-  BsHouseDoor,
-  BsLaptop,
-  BsPerson,
   BsPersonAdd,
 } from "react-icons/bs";
-import ComponentCard from "../../components/common/ComponentCard";
-import Label from "../../components/form/Label";
-import Input from "../../components/form/input/InputField";
+import ComponentCard from "../../../components/common/ComponentCard";
+import Label from "../../../components/form/Label";
+import Input from "../../../components/form/input/InputField";
+import { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { fetchMembersByLeader } from "../../../features/membersByLeader/membersByLeaderSlice";
+import { workDistribution } from "../../../features/workDistribution/workDistributionSlice";
 
 export default function WorkGroup() {
+  const members = useSelector((state) => state.membersByLeader.items);
+  const dispatch = useDispatch();
+  const today = new Date().toISOString().split("T")[0];
+
+  const [formData, setFormData] = useState({
+    work_title: "",
+    work_desc: "",
+    work_expire_date: "",
+    work_date: today,
+    work_emp_id: "",
+  });
+
+  function handleSubmit(e) {
+    e.preventDefault();
+    dispatch(workDistribution(formData));
+  }
+
+  const handleChange = (e) => {
+    const { name } = e.target;
+    let { value } = e.target;
+    if (name === "work_emp_id") value = value === "" ? "" : Number(value);
+    setFormData((prevData) => ({
+      ...prevData,
+      [name]: value,
+    }));
+  };
+
+  useEffect(() => {
+    if (members.length <= 0) {
+      dispatch(fetchMembersByLeader());
+    }
+    if (members.length > 0 && !formData.work_emp_id) {
+      setFormData((prev) => ({ ...prev, work_emp_id: members[0].empId }));
+    }
+  }, [members, dispatch, formData.work_emp_id]);
+
   return (
     <ComponentCard title="Input Group">
       <div className="grid grid-cols-1  gap-6">
@@ -31,6 +58,9 @@ export default function WorkGroup() {
           <div className="relative">
             <Input
               placeholder="Enter Task Title"
+              onChange={handleChange}
+              name="work_title"
+              value={formData.work_title}
               type="text"
               className="pl-[62px]"
             />
@@ -46,6 +76,10 @@ export default function WorkGroup() {
           <div className="relative">
             <textarea
               placeholder="Enter Task Description"
+              onChange={handleChange}
+              name="work_desc"
+              value={formData.work_desc}
+              type="text"
               rows={3}
               className="w-full rounded-md border border-gray-300 bg-white pl-[62px] pr-3 py-3 text-gray-700 focus:outline-none dark:border-gray-800 dark:bg-gray-900 dark:text-gray-300"
             />
@@ -59,7 +93,13 @@ export default function WorkGroup() {
         <div>
           <Label>Expire Date</Label>
           <div className="relative">
-            <Input type="date" className="pl-[62px]" />
+            <Input
+              type="date"
+              className="pl-[62px]"
+              onChange={handleChange}
+              name="work_expire_date"
+              value={formData.work_expire_date}
+            />
             <span className="absolute left-0 top-1/2 -translate-y-1/2 border-r border-gray-200 px-3.5 py-3 text-gray-500 dark:border-gray-800 dark:text-gray-400">
               <BsCalendar2Event className="!w-6 !h-6 text-gray-500" />
             </span>
@@ -70,11 +110,20 @@ export default function WorkGroup() {
         <div>
           <Label>User Name</Label>
           <div className="relative">
-            <select className="w-full rounded-md border border-gray-300 bg-white pl-[62px] py-3 text-gray-700 focus:outline-none dark:border-gray-800 dark:bg-gray-900 dark:text-gray-300">
-              <option value="">Select User</option>
-              <option value="user1">John Doe</option>
-              <option value="user2">Jane Smith</option>
-              <option value="user3">Robert Brown</option>
+            <select
+              onChange={handleChange}
+              name="work_emp_id"
+              value={formData.work_emp_id}
+              className="w-full rounded-md border border-gray-300 bg-white pl-[62px] py-3 text-gray-700 focus:outline-none dark:border-gray-800 dark:bg-gray-900 dark:text-gray-300"
+            >
+              <option value="" disabled>
+                {members.length ? "Select member" : "No members available"}
+              </option>
+              {members.map((member) => (
+                <option key={member.empId} value={member.empId}>
+                  {member.empName}
+                </option>
+              ))}
             </select>
             <span className="absolute left-0 top-1/2 -translate-y-1/2 border-r border-gray-200 px-3.5 py-3 text-gray-500 dark:border-gray-800 dark:text-gray-400">
               <BsPersonAdd className="!w-6 !h-6 text-gray-500" />
@@ -86,6 +135,7 @@ export default function WorkGroup() {
         <div className="pt-4 flex items-center justify-center">
           <button
             type="submit"
+            onClick={handleSubmit}
             className="w-1/3 rounded-xl bg-green-600 px-6 py-3 font-semibold text-white shadow-md transition hover:bg-green-700"
           >
             Submit
